requirements.txt
fastapi==0.115.12
httpx==0.27.2
python-multipart==0.0.9
python-dotenv==1.0.1
uvicorn[standard]==0.30.6
Flask==2.3.0
flask-cors==4.0.0
Werkzeug<3
